requirements.txt
pandas
numpy
pyarrow
scikit-learn
lightgbm
statsmodels
paho-mqtt
pyyaml
joblib
